Add routes folder for Router
