+++ src/index.tsx
@@ -1,27 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import {
-  createBrowserRouter,
-  RouterProvider,
-  Navigate,
-} from "react-router-dom";
-import SharedPage from "./pages/SharedPage";
-import FolderPage from "./pages/FolderPage";
-
-const router = createBrowserRouter([
-  {
-    path: "/",
-    element: <Navigate to="/shared" replace />,
-  },
-  {
-    path: "/shared",
-    element: <SharedPage />,
-  },
-  {
-    path: "/folder",
-    element: <FolderPage />,
-  },
-]);
+import { RouterProvider } from "react-router-dom";
+import router from "routes/Router";
 
 const root = document.getElementById("root") as HTMLElement;
